Combined all the parts into a single (python) I3Module. Also: lots of minor improvements.
git-svn-id: http://code.icecube.wisc.edu/svn/sandbox/ckopper/clsim@72771 16731396-06f5-0310-8873-f7f720988828

+++ resources/kernels/propagation_kernel.cl
@@ -46,7 +46,7 @@ struct I3CLSimStep
     uint numPhotons;
     float weight;
     uint identifier;
-    float dummy;
+    uint dummy;
 };
 
 struct I3CLSimPhoton 
@@ -58,7 +58,7 @@ struct I3CLSimPhoton
     uint numScatters; // number of scatters
     float weight;
     uint identifier;
-    float dummy;
+    uint dummy;
 };
